This changes implements the core functions for AES-GCM stitched encryption/decryption with XMM/PCLMULQDQ/AES-NI in Rust. The implementation provides a safe Rust interface that encapsulates the use of unsafe intrinsics and pointers. It also implements the C FFI functions `SymCryptAesGcmEncryptStitchedXmm` and `SymCryptAesGcmDecryptStitchedXmm`.

To limit the size and churn of this PR, I'm continuing with the next parts of the implementation in another working branch. Subsequent PRs will implement:
- End-to-end AES GCM functionality (i.e. `SymCryptGcmEncrypt`/`SymCryptGcmDecrypt`)
- NEON implementations

+++ lib/rsa_padding.c
@@ -297,7 +297,7 @@ SymCryptRsaPkcs1RemoveEncryptionPadding(
     mPaddingError |= SymCryptMask32LtU31( iFirstZero, 10 );
 
     // Compute the # bytes of the message; 0 if there was a padding error
-    cbPlaintextResult = ~mPaddingError & (cbPkcs1Format - iFirstZero - 1);
+    cbPlaintextResult = ~mPaddingError & ((UINT32) cbPkcs1Format - iFirstZero - 1);
 
     // We're done if the caller didn't want the actual message, but only the size.
     // We do that before checking the size of the plaintext buffer so that callers who
